+++ src/main/kotlin/com/tistory/shanepark/dutypark/security/domain/dto/LoginMember.kt
@@ -4,7 +4,6 @@ data class LoginMember(
     val id: Long,
     val email: String? = null,
     val name: String,
-    val departmentId: Long? = null,
     val department: String? = null,
     var isAdmin: Boolean = false
 ) {
