+++ src/main/resources/templates/team/myteam.html
@@ -1,7 +1,174 @@
-<div>
-  MY TEAM
+<div id="app" class="card p-0 rounded shadow-sm my-team">
+  <div v-if="!team">
+    <div class="card-header bg-secondary text-white fw-bold mb-1 fs-2">내 팀</div>
+    <div class="justify-content-center align-items-center p-5">
+      <p class="fs-3 fw-bold ms-2 text-align-center"><i class="bi bi-person-circle fs-1 me-3"></i>어느 팀에도 속해있지 않습니다.</p>
+      <p class="fs-3 fw-bold ms-2 text-align-center">팀 관리자에게 가입을 요청해주세요.</p>
+    </div>
+  </div>
+  <template v-if="team">
+    <div class="card-header text-center bg-secondary text-white fw-bold mb-1 fs-2" v-text="team.name"></div>
+    <div class="d-flex justify-content-between align-items-center mb-3 p-1">
+      <button v-on:click="addMonth(-1)" class="fs-2 btn btn-outline-dark">&lt;</button>
+      <h2 class="fs-2 fw-bold mb-0 text-center flex-grow-1" v-text="year + '년 '+ month +'월'"></h2>
+      <button v-on:click="addMonth(1)" class="fs-2 btn btn-outline-dark">&gt;</button>
+      <button v-on:click="goToToday" class="fs-2 btn btn-outline-primary ms-2">오늘</button>
+    </div>
+
+    <!--  Calendar-->
+    <div class="row row-cols-7 fw-bold mx-0 fs-2">
+      <div v-for="(day, index) in weekDays"
+           class="col border border-2 text-center"
+           :class="{ red: index === 0, blue: index === 6 }"
+           v-text="day">
+      </div>
+      <div v-for="(teamDay, index) in teamDays"
+           @click="selectDay(teamDay)"
+           class="col small p-2 border cursor-pointer hover-bg-light"
+           :class="['BACKGROUND-' + myDuty[index]?.dutyColor, {'today': isToday(teamDay), 'border-dark border-3': isSelectedDay(teamDay) }]"
+      >
+        <div :class=" {'red': index%7==0, 'blue': index%7==6, 'opacity-50': teamDay.month!=month }">
+          <div class="fw-bold day">{{teamDay.day}}
+          </div>
+          <!--        This is for when the Team Schedule feature is added-->
+          <!--        <div v-if="index%3==0" class="small text-truncate">회의</div>-->
+        </div>
+      </div>
+    </div>
+    <!-- End Calendar-->
+
+    <!--  shift-->
+    <div class="mt-2 p-3" v-if="shift.length > 0">
+      <h2 class="fs-3 mb-0 text-right flex-grow-1 mb-2"
+          v-text="selDay.year + '년 '+ selDay.month +'월 ' + selDay.day + '일'"></h2>
+      <template v-for="group in shift"
+                v-if="group.members.length>0">
+        <div
+            :class="{'duty-type border rounded p-0 mb-4 overflow-hidden': true, 'border-dark border-2': group.isMyGroup}"
+            v-if="group.members.length > 0">
+          <div
+              :class="['BACKGROUND-' + group.dutyType.color, 'p-3', 'd-flex', 'justify-content-between', 'align-items-center']">
+            <span class="fw-bolder fs-3" v-text="group.dutyType.name"></span>
+            <span class="badge bg-light text-dark fs-3">{{ group.members.length }}</span>
+          </div>
+          <div class="m-0 row p-3">
+            <div v-for="member in group.members" :key="member.id" class="col-3"
+                 @click.prevent="moveTo(member.id)">
+              <div
+                  class="member-card p-3 border rounded shadow-sm text-center cursor-pointer fs-3"
+                  :class="{'border-dark border-3': member.id === loginMemberId}"
+              >
+                <i class="bi bi-person-fill mb-2"></i>
+                <span v-text="member.name"></span>
+              </div>
+            </div>
+          </div>
+        </div>
+      </template>
+    </div>
+    <!--  End Shift-->
+  </template>
 </div>
 
 <script>
-
+    const today = new Date();
+    const app = new Vue({
+        el: '#app',
+        data: {
+            loginMemberId: parseInt("[[${loginMember?.id}]]"),
+            today: {
+                year: today.getFullYear(),
+                month: today.getMonth() + 1,
+                day: today.getDate()
+            },
+            selDay: {
+                year: today.getFullYear(),
+                month: today.getMonth() + 1,
+                day: today.getDate()
+            },
+            year: today.getFullYear(),
+            month: today.getMonth() + 1,
+            day: today.getDate(),
+            teamDays: [],
+            weekDays: ['일', '월', '화', '수', '목', '금', '토'],
+            team: {
+                name: ''
+            },
+            shift: [],
+            myDuty: [],
+        },
+        mounted() {
+            this.loadMyTeam();
+            this.loadShift();
+            this.loadMyDuty();
+        },
+        computed: {},
+        methods: {
+            loadMyTeam() {
+                fetch('/api/departments/my?year=' + this.year + '&month=' + this.month)
+                    .then(response => response.json())
+                    .then(data => {
+                        app.team = data.department;
+                        app.teamDays = data.teamDays;
+                    })
+                    .catch(error => {
+                        console.error('Error:', error);
+                    });
+            },
+            addMonth(month) {
+                const newDate = new Date(app.year, app.month - 1 + month);
+                app.year = newDate.getFullYear();
+                app.month = newDate.getMonth() + 1;
+                app.loadMyTeam();
+                app.loadMyDuty();
+            },
+            isToday(teamDay) {
+                return teamDay.year === this.today.year &&
+                    teamDay.month === this.today.month &&
+                    teamDay.day === this.today.day;
+            },
+            loadShift() {
+                fetch(`/api/departments/shift?year=${this.selDay.year}&month=${this.selDay.month}&day=${this.selDay.day}`)
+                    .then(response => response.json())
+                    .then(data => {
+                        app.shift = data;
+                        app.shift.forEach(group => {
+                            group.isMyGroup = group.members.some(member => member.id === app.loginMemberId);
+                        });
+                    })
+                    .catch(error => {
+                        console.error('Error:', error);
+                    });
+            },
+            selectDay(teamDay) {
+                app.selDay = teamDay;
+                app.loadShift();
+            },
+            isSelectedDay(teamDay) {
+                return teamDay.year === this.selDay.year &&
+                    teamDay.month === this.selDay.month &&
+                    teamDay.day === this.selDay.day;
+            },
+            moveTo(id) {
+                location.href = "/duty/" + id;
+            },
+            goToToday() {
+                this.year = this.today.year;
+                this.month = this.today.month;
+                this.selDay = {...this.today};
+                this.loadMyTeam();
+                this.loadShift();
+            },
+            loadMyDuty() {
+                fetch(`/api/duty?year=${this.year}&month=${this.month}&memberId=${this.loginMemberId}`)
+                    .then(response => response.json())
+                    .then(data => {
+                        app.myDuty = data;
+                    })
+                    .catch(error => {
+                        console.error('Error:', error);
+                    });
+            },
+        },
+    });
 </script>
